Add files via upload
Training of the model used by MitoSplit-Net. Input data is of mitochondria from a SIM microscope, while the expected output is a map of division events. The latter was generated by Dora Mahecic, who manually annotated the output of a Hessian Filter applied to the mitochondria images plus Drp1 SIM data.