Change delFissions for selFissions
Previous approach for spatiotemporal filtering of events was based on removing areas with low event score. The problem is that we also want to filter out events that last less than 3 frames, and this is not taken into account by simply thresholding the event score signal per fission site. A solution to this issue was to generate a list of labels that satisfied all of these criteria and add them iteratively to the final processed output and label data.